pyproject.toml
[project]
name = "soundnestgrpc"
version = "0.1.0"
description = ""
authors = [
    {name = "Navi",email = "168604098+King-Navi@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11.9,<4.0"
dependencies = [
    "grpcio (>=1.71.0,<2.0.0)",
    "protobuf (>=5.26,<6.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "sqlacodegen (>=3.0.0,<4.0.0)",
    "cryptography (>=44.0.3,<45.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
]

[tool.poetry]
packages = [{include = "soundnestgrpc", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

#Pylint
[tool.pylint.main]
disable = [
    "missing-module-docstring", 
    "missing-class-docstring", 
    "missing-function-docstring"
]
[tool.pylint.format]
max-line-length = 200